Добавлен интерфейс IUserRepository и класс UserRepository, реализующий связь с таблицей users. Создан отедльный контроллер для действий с пользователями.
