+++ StagingWizard/StagingWizard/DataLayerContracts/IUserRepository.cs
@@ -1,0 +1,15 @@
+﻿using System;
+using StagingWizard.UIContracts;
+using System.Collections.Generic;
+using System.Linq;
+using System.Threading.Tasks;
+
+namespace StagingWizard.DataLayerContracts
+{
+    public interface IUserRepository
+    {
+        string SignInCheck(User user);
+        string AddUser(User user);
+        bool CheckToken(string token);
+    }
+}
